+++ src/AdapterAzureBlob.ts
@@ -117,6 +117,9 @@ export class AdapterAzureBlob extends AbstractAdapter {
         this._configError = `[configError] ${e.message}`;
       }
     }
+    if (typeof this.config.bucketName !== "undefined") {
+      this._bucketName = this.config.bucketName;
+    }
   }
 
   // protected, called by methods of public API via AbstractAdapter