+++ publish/AdapterGoogleCloud/changelog.md
@@ -1,4 +1,8 @@
-1.0.3
+# 1.0.4
+
+- fix bug in selected bucket
+
+# 1.0.3
 
 - revert to v1 format of config URLs
 - re-implement storing the selected bucket in local state